% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_scan.R
\name{analyze_parameter_scan}
\alias{analyze_parameter_scan}
\title{Analyze results of a parameter scan}
\usage{
analyze_parameter_scan(parameter_scan_results)
}
\arguments{
\item{parameter_scan_results}{String or List. If a string, it is
interpreted as the name of a \code{rds} file that contains the results of a
parameter scan which is then loaded using \code{\link[=readRDS]{readRDS()}}. Otherwise, it
should be the output of \code{\link[=run_parameter_scan]{run_parameter_scan()}} directly.}
}
\description{
Analyze results of a parameter scan
}
\seealso{
\code{\link[=run_parameter_scan]{run_parameter_scan()}}, \code{\link[=readRDS]{readRDS()}}
}
