% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append_to_table}
\alias{append_to_table}
\title{Write *data* to supplied file in append mode without generating a warning 
message.}
\usage{
append_to_table(data, filename, ...)
}
\arguments{
\item{data}{Any object which can be handled by `write.table`.}

\item{filename}{Name of file to append to.}

\item{...}{All additional arguments are passed to `write.table`.}
}
\description{
This function essentially wraps `write.table` with a calling handler that
suppresses appending warnings that would appear with the argument 
`col.names = TRUE`.
}
