% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{box_smooth}
\alias{box_smooth}
\title{Endpoint smoother}
\usage{
box_smooth(x, box_width = 28)
}
\arguments{
\item{x}{1D data to be smoothed.}

\item{box_width}{Width (in units of vector steps) of the box used for 
smoothing.}
}
\value{
x_smooth Smoothened version of *x*.
}
\description{
Smooth data in vector *x* to its endpoint.
}
\details{
Employ an endpoint box filter (aka "running mean" or midpoint smoother) to the 
1-D data in *x*:
`x_smoothed[i] = mean(x[i-box_width:i])`
Where *x* is considered to be *zero-padded* vor values of *i-box_width* < 1.
}
