% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\name{build_functional_group}
\alias{build_functional_group}
\title{Build the effective functional group as a weighted linear combination.}
\usage{
build_functional_group(P)
}
\arguments{
\item{P}{list; name-value pairs of parameters. Has to contain w_FGX with X 
in [A, B, C, D].}
}
\value{
A modified list, where all functional group parameters are updated 
  to the values of the effective functional group according to the weights 
  w_FGX.
}
\description{
Uses the weights found in :param:`P` to construct the effective functional 
groups and updates functional group parameters in P.
}
