% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cCO2_coefficients}
\alias{cCO2_coefficients}
\title{Coefficients for polynomial fits to atmospheric CO2 concentration scenarios.}
\format{
A data.frame with columns
\describe{
\item{RCP}{Name of the RCP.}
\item{a3}{Coefficient of the quadratic term.}
\item{a2}{Coefficient of the quadratic term.}
\item{a1}{Coefficient of the linear term.}
\item{a0}{Coefficient of the constant term.}
}
}
\usage{
cCO2_coefficients
}
\description{
The fifth IPCC assessment report provides data for atmospheric CO2
concentration under different representative concentration pathway (RCP)
scenarios on decadal resolution. The values in this matrix represent the
coefficients of a second order polynomial fit to said data.
}
\details{
These coefficients can be used to produce functions of the form \code{f(year) =  a3*year^3 + a2*year^2 + a1*year + a0} which return the estimated
atmospheric CO2 concentration in ppm for a given year.
The validity of the polynomial fit is restricted to the years 2000 to 2100.
}
\seealso{
\code{\link[=atmospheric_CO2]{atmospheric_CO2()}}, \code{\link[=WeatherData]{WeatherData()}}\verb{$aCO2()}
}
\keyword{datasets}
