% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinator.R
\name{create_combinations}
\alias{create_combinations}
\title{Create Valid Combinations}
\usage{
create_combinations(param_values)
}
\arguments{
\item{param_values}{A list giving all options for the parameter values
which are to be combined. The format is \code{list[[param_name]] =  param_values} where \code{param_values} is a vector with the values for the
respective parameter. The parameter names for functional group weights
(\code{w_FGX} with \code{X} in (A, B, C, D)) receive special treatment and
therefore need to be spelled correctly.}
}
\value{
combinations An unnamed list where every entry is a list
containing the parameter values (named as in the input \emph{param_values})
for a valid combination.
}
\description{
Generate a list which contains all possible combinations of the provided
parameter values. This excludes combinations that are invalid because the
sum criterion for functional groups \code{w_FGA + w_FGB + w_FGC + w_FGD = 1} is
not fulfilled.
}
\details{
Assume for example the following list as argument \emph{param_values}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(w_FGA = c(0, 0.5, 1), w_FGB = c(0, 0.5, 1), NI = c(0.5, 0.9))
}\if{html}{\out{</div>}}

This would generate the combinations\tabular{lll}{
   w_FGA \tab w_FGB \tab NI \cr
   0 \tab 1 \tab 0.5 \cr
   0 \tab 1 \tab 0.9 \cr
   0.5 \tab 0.5 \tab 0.5 \cr
   0.5 \tab 0.5 \tab 0.9 \cr
   1 \tab 0 \tab 0.5 \cr
   1 \tab 0 \tab 0.9 \cr
}
}
