% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{create_example_environment}
\alias{create_example_environment}
\title{Provide an example ModvegeEnvironment}
\usage{
create_example_environment(site = "posieux")
}
\arguments{
\item{site}{Choose for which example site an environment is to be created.
Options: \code{"posieux"}, \code{"sorens"}.}
}
\value{
E A \code{ModvegeEnvironment} instance based on the example data for
\emph{site} which is shipped with this package.
}
\description{
This is intended for testing and for the examples in the documentation.
}
\examples{
extdata = system.file("extdata", package = "growR") 
print(extdata)
list.files(extdata, recursive = TRUE)
create_example_environment()

}
