% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{DEBUG_LEVELS}
\alias{DEBUG_LEVELS}
\title{Names of debug levels.}
\format{
An object of class \code{character} of length 5.
}
\usage{
DEBUG_LEVELS
}
\description{
Names of debug levels.
}
\keyword{datasets}
