% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enforce_mono_inc}
\alias{enforce_mono_inc}
\title{enforce monotonically increasing}
\usage{
enforce_mono_inc(y)
}
\arguments{
\item{y}{values to process}
}
\description{
enforce monotonically increasing
}
