% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ensure_table_columns}
\alias{ensure_table_columns}
\title{Check if supplied table contains all *required* variables.}
\usage{
ensure_table_columns(required, data, data_name = "the data table")
}
\arguments{
\item{required}{List of names of required variables.}

\item{data}{data.frame or similar object to be checked.}

\item{data_name}{Name to be displayed in the error message if any variable 
is missing.}
}
\description{
Logs an error if any variable is missing and lists the missing variables 
in the error message along with *data_name*.
}
