% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{ERROR}
\alias{ERROR}
\title{Constants representing different debug levels to be used internally.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ERROR
}
\description{
Constants representing different debug levels to be used internally.
}
\keyword{datasets}
