% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{fCO2_transpiration_mod}
\alias{fCO2_transpiration_mod}
\title{CO2 transpiration modifier}
\usage{
fCO2_transpiration_mod(c_CO2, c_ref = 360)
}
\arguments{
\item{c_CO2}{numeric Atmospheric CO2 concentration in ppm}

\item{c_ref}{numeric Reference CO2 concentration in ppm.}
}
\description{
Function describing the effects of elevated CO2 on transpiration.
}
\details{
The function for the effect on transpiration is from equations (2-6) in:

Kruijt, B., J.-P. M. Witte, C. M. J. Jacobs and T. Kroon, 2008:
   Effects of rising atmospheric CO2 on evapotranspiration and soil
   moisture: A practical approach for the Netherlands. J. Hydrol.,
   349, 257-267.

It appears that this paper that said equations are most likely incorrect. 
With the stated values, I cannot reproduce tabulated values of c close to 
1, as in their table 3. Instead,I conclude that equation (4) should read::

  c = 1 + s_gs * s_T * F_T * deltaCO2

with the multiplicative terms giving small negative numbers.
The factors s_gs, s_T and F_T for grasslands are taken from pages 260 and 
261 in Kruijt et al. where we averaged over the stated ranges to get:
c ~= 1 + 0.0001 * deltaCO2
}
