% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\docType{data}
\name{FG_A}
\alias{FG_A}
\title{Functional group A}
\format{
An object of class \code{FunctionalGroup} (inherits from \code{R6}) of length 23.
}
\usage{
FG_A
}
\description{
Functional group A
}
\seealso{
[FunctionalGroup]
}
\keyword{datasets}
