% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\docType{data}
\name{FG_B}
\alias{FG_B}
\title{Functional group B}
\format{
An object of class \code{FunctionalGroup} (inherits from \code{R6}) of length 23.
}
\usage{
FG_B
}
\description{
Functional group B
}
\seealso{
[FunctionalGroup]
}
\keyword{datasets}
