% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\docType{data}
\name{FG_C}
\alias{FG_C}
\title{Functional group C}
\format{
An object of class \code{FunctionalGroup} (inherits from \code{R6}) of length 23.
}
\usage{
FG_C
}
\description{
Functional group C
}
\seealso{
[FunctionalGroup]
}
\keyword{datasets}
