% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\docType{data}
\name{FG_D}
\alias{FG_D}
\title{Functional group D}
\format{
An object of class \code{FunctionalGroup} (inherits from \code{R6}) of length 23.
}
\usage{
FG_D
}
\description{
Functional group D
}
\seealso{
[FunctionalGroup]
}
\keyword{datasets}
