% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{fPAR}
\alias{fPAR}
\title{Radiation limitation}
\usage{
fPAR(PAR)
}
\arguments{
\item{PAR}{float Photosynthetically active radiation in MJ/m^2}
}
\value{
A value in the range [0, 1], acting as a multiplicative factor to 
  plant growth.
}
\description{
Threshold function representing growth limitation due to lack of 
photosynthetically active radiation (PAR).
}
\examples{
fPAR(4)

}
