% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\name{FunctionalGroup}
\alias{FunctionalGroup}
\title{Representation of a grassland plant population}
\description{
A functional group is a representation of a grassland plant population 
with certain functional attributes.

It contains many plant parameters that are collected under the hood of
functional groups. The class implements S3 style operator overloading such 
that one can do things like 
```
mixed_functional_group = 0.8 * FG_A + 0.2 * FG_B
```
}
\details{
Default values for parameters are taken from functional group A in
Jouven et al.
}
\references{
{
 \insertRef{jouven2006ModelPredictingDynamics}{rmodvege}
}
}
\seealso{
[FunctionalGroup$set_parameters()]
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FunctionalGroup-new}{\code{FunctionalGroup$new()}}
\item \href{#method-FunctionalGroup-get_parameters}{\code{FunctionalGroup$get_parameters()}}
\item \href{#method-FunctionalGroup-get_parameters_ordered}{\code{FunctionalGroup$get_parameters_ordered()}}
\item \href{#method-FunctionalGroup-set_parameters}{\code{FunctionalGroup$set_parameters()}}
\item \href{#method-FunctionalGroup-set_parameters_ordered}{\code{FunctionalGroup$set_parameters_ordered()}}
\item \href{#method-FunctionalGroup-clone}{\code{FunctionalGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionalGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionalGroup-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionalGroup$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Key-value pairs of parameters to be set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionalGroup-get_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionalGroup-get_parameters}{}}}
\subsection{Method \code{get_parameters()}}{
Convenient getter

Returns all parameters with their names in a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionalGroup$get_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionalGroup-get_parameters_ordered"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionalGroup-get_parameters_ordered}{}}}
\subsection{Method \code{get_parameters_ordered()}}{
Ordered getter

Returns all parameters in reproducible order in a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionalGroup$get_parameters_ordered()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionalGroup-set_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionalGroup-set_parameters}{}}}
\subsection{Method \code{set_parameters()}}{
Convenient setter

Set all specified parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionalGroup$set_parameters(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Key-value pairs of parameters to be set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionalGroup-set_parameters_ordered"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionalGroup-set_parameters_ordered}{}}}
\subsection{Method \code{set_parameters_ordered()}}{
Efficient setter, assumes parameters come in known order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionalGroup$set_parameters_ordered(ordered_parameter_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ordered_parameter_values}}{Parameter values to be set. Need to be 
in the same order as `FunctionalGroup$fg_parameter_names`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionalGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionalGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionalGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
