% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gc_logistic}
\alias{gc_logistic}
\title{fit logistic with parameter description as per growthcurver package}
\usage{
gc_logistic(t, k, n0, r)
}
\arguments{
\item{t}{time}

\item{k}{the maximum possible population size in a particular environment, or the carrying capacity}

\item{n0}{the population size at the beginning of the growth curve}

\item{r}{The intrinsic growth rate of the population, r, is the growth rate that would occur if there were no restrictions imposed on total population size.}
}
\value{
y-values of the specified model for given time points.
}
\description{
fit logistic with parameter description as per growthcurver package
}
\examples{
gc_logistic(1:95, 10, .1, .2)
}
