% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{get_expected_n_cuts}
\alias{get_expected_n_cuts}
\title{Return the number of expected cuts for a site at a given *elevation* and 
management *intensity*.}
\usage{
get_expected_n_cuts(elevation, intensity = "high")
}
\arguments{
\item{elevation}{The elevation of the considered site in meters above sea 
level.}

\item{intensity}{One of ["high", "middle", "low", "extensive". Management 
intensity for considered site.}
}
\value{
Number of expected cuts per season.
}
\description{
This uses :data.frame:`management_parameters` as a lookup table and 
interpolates linearly in between the specified values.
}
