% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionalData.R
\name{getRegionalData}
\alias{getRegionalData}
\title{getRegionalData}
\usage{
getRegionalData(Longitude, Latitude, realm = c("Marine", "Terrestrial",
  "Freshwater", "IPBES", "Hotspot"))
}
\arguments{
\item{realm}{"Marine", "Terrestrial", "Freshwater", "IPBES", "Hotspot"}

\item{lats}{Latitude values in decimal degrees}

\item{longs}{Longitude values in decimal degrees}
}
\value{
Returns a data frame of geographic regions for each point

seealso \code{\link{sp::over}}
}
\description{
\code{getRegions} Gets regional information for your dataset, depending on the realm of interest.
#' Code is modified and extended from @jebyrnes package meowR {\link{https://github.com/jebyrnes/meowR}}.
}
\examples{

data(regions)

latlong <- data.frame(lat = c(50.0944444, 33.4691667, 34.0348833, 
     55.0869876, 51.7787776, 
     49.6088889, -35.750729),
     long=c(-127.55889, -119.53028, 
     -119.70308, -161.87444, 
     178.30199, -126.78056, 150.259155))
     
regionalData <- getRegionalData(latlong$lat, latlong$long, realm="Marine")

newdata <- cbind(latlong, regionalData)

newdata
}
\author{
Sarah Supp, Shane Blowes.
}
