% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gf_gompertz}
\alias{gf_gompertz}
\alias{gf_richards}
\alias{gf_logistic}
\alias{gf_gompertz_exp}
\alias{gf_spline}
\title{Compute the output of a model from the time and specified parameters.}
\usage{
gf_gompertz(time, A, mu, lambda)

gf_richards(time, A, mu, lambda, nu = 0.1)

gf_logistic(time, A, mu, lambda)

gf_gompertz_exp(time, A, mu, lambda, alpha = 0.1,
  tshift = max(time)/10)

gf_spline(time, n)
}
\arguments{
\item{time}{\code{<num>} Time points (x-axes) for which the function values will be returned.}

\item{A}{\code{<num>} Maximum of the curve.}

\item{mu}{\code{<num>} Maximum slope.}

\item{lambda}{\code{<num>} Lag phase}

\item{nu}{\code{<num>} shape parameter, enables flexible adjustment that the point of inflexion can be at any value between zero and A.}

\item{alpha}{\code{<num>} control the strength (slope) of the second increase.}

\item{tshift}{\code{<num>} control the location (time) of the second increase after the function enters a first saturation plateau}

\item{n}{\code{<num>} number of cells or proxy for number of cells}
}
\value{
y-values of the specified model for given time points.
}
\description{
Compute the output of a model from the time and specified parameters.
}
\details{
Gompertz model parametrized by \code{A}, \code{mu}, \code{lambda}:
}
