% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growR-package.R
\docType{data}
\name{growR_package_options}
\alias{growR_package_options}
\title{Default options introduced by package growR}
\format{
An object of class \code{list} of length 4.
}
\usage{
growR_package_options
}
\description{
These are the default options, set when the package is loaded 
  by `library(growR)`. To get or change the current value of an option, 
use the `options()` function.

\describe{
  \item{growR.verbosity}{Integer that controls how much console output is 
        generated by growR functions. Higher numbers mean more output. 
        See [logger()].}
  \item{growR.input_dir}{Name of the directory in which to look for 
        input files.}
  \item{growR.output_dir}{Name of the directory into which output files
        are written.}
  \item{growR.data_dir}{Name of the directory in which to look for 
        measured data files.}
}
}
\seealso{
[options()]
}
\keyword{datasets}
