% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{initial_condition_values}
\alias{initial_condition_values}
\title{ICV}
\format{
An object of class \code{list} of length 0.
}
\usage{
initial_condition_values
}
\description{
ICV
}
\keyword{datasets}
