% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{is_todo}
\alias{is_todo}
\title{check if a value indicates there is something to do.}
\usage{
is_todo(x)
}
\arguments{
\item{x}{value to check if is to do.}
}
\value{
\link{TRUE} or \link{FALSE} indicating if there is something to do or not.
}
\description{
Similar to a 'true-ish' type value. Everything but \link{FALSE} and \link{NA} should
return \link{TRUE}.
}
\examples{
is_todo(NA)
is_todo(3)
is_todo('a')
is_todo(TRUE)
is_todo(FALSE)
is_todo(I)
is_todo(function(y) {y})
}
\keyword{internal}
