% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{load_measured_data}
\alias{load_measured_data}
\title{Load experimental data}
\usage{
load_measured_data(filenames)
}
\arguments{
\item{filenames}{Vector of strings representing simulation output 
filenames.}
}
\value{
measured_data list of data.frame each corresponding to one of 
 the sites detected in *filenames*. Each data.frame contains the keys
 - dBM
 - cBM
 - year
 - DOY
}
\description{
Load all datasets associated with any site present in the names of the 
supplied list of files.
}
\details{
This function internally uses [get_site_name] and makes the same 
assumptions about the output filename formats. It further assumes 
measured data to be located in "data/" and adhere to the filename 
format "x.csv" with *x* being the site name.

Upon loading, the cumulative biomass growth *cBM* is automatically 
calculated from the given daily biomass growth *dBM* values.
}
