% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logger}
\alias{logger}
\title{Primitive logger for debugging.}
\usage{
logger(msg = "", level = DEBUG, stop_on_error = TRUE)
}
\arguments{
\item{msg}{The message to print.}

\item{level}{The message will only be printed if its *level* is lower 
or equal to *DEBUG_LEVEL*}

\item{stop_on_error}{Can be set to FALSE in order to continue 
execution despite emitting a message of *level* ERROR.}
}
\description{
Primitive logger for debugging.
}
