% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_yearDOY}
\alias{make_yearDOY}
\title{Create unique DOY + year identifier}
\usage{
make_yearDOY(years, DOYs)
}
\arguments{
\item{years}{int vector.}

\item{DOYs}{int vector of same length as *years*.}
}
\value{
int vector of same length as *years* containing numbers of the 
  form YYYYDDD, where the first four digits represent a year and the last 
  four represent a DOY.
}
\description{
Return numbers of the form YYYYDDD where YYYY is the year and DDD the DOY.
}
