% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGROWmapData.R
\name{makeGROWmapData}
\alias{makeGROWmapData}
\title{makeMEOWmapData}
\usage{
makeGROWmapData(newdata, fillColName, fillAlphaColName = NULL,
  regionColName = type, type = "ECOREGION", realm = "Marine",
  excludeNoDataAreas = T)
}
\arguments{
\item{newdata}{dataset to be used}

\item{fillColName}{name of the column with information to be used for a fill value}

\item{fillAlphaColName}{Colname of column used to scale alpha level of fill}

\item{regionColName}{data name of column with ecoregion, province, or realm names}

\item{type}{are these values from an ECOREGION, PROVINCE, or REALM map. Defaults to "ECOREGION". Note all caps.}

\item{realm}{What geographic realm are the data from (Marine, Freshwater, Terrestrial, Hotspot, IPBES)?}

\item{excludeNoDataAreas}{Exclude areas/regions from the plot with no data?}
}
\value{
Returns \code{\link{data.frame}} object
}
\description{
\code{makeGROWmapData} takes a dataset with values and a GROW column merges them for use by \code{\link{makeGROWmap}}. 
Code is modified and extended from @jebyrnes package meowR {\link{https://github.com/jebyrnes/meowR}}.
}
\details{
\code{makeGROWmapData} takes a dataset with values and a GROW column and makes a joined \code{\link{data.frame}}
}
\examples{

data(regions.df)

ndf <- data.frame(Ecoregions = levels(regions.df$ECOREGION), 
Values = runif(length(levels(regions.df$ECOREGION)), 0,100))

head(makeGROWmapData(ndf, fillColName="Values", regionColName="Ecoregions"))

}
\author{
Sarah Supp
}
