% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_auc}
\alias{metric_auc}
\title{Area under the curve (integral of fit over range)}
\usage{
metric_auc(fit, x, ...)
}
\arguments{
\item{fit}{model object with fitted values}

\item{limits}{<num> vector of length; lower and upper limits of integration}
}
\description{
Area under the curve (integral of fit over range)
}
