% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_avg_dbl}
\alias{metric_avg_dbl}
\title{Average Doubling or Generation Time from start until max rate.}
\usage{
metric_avg_dbl(x, y, dy, ...)
}
\description{
x,y coordinates of first point of maximum dy. slope (m), y intercept (b),
x intercept (y0_x) of tangent line
}
\details{
Finds the point of maximum growth rate then treating that as the end point,
computes the average
}
\keyword{internal}
