% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_coefs}
\alias{metric_coefs}
\title{coefficients of a model fit}
\usage{
metric_coefs(fit, ...)
}
\arguments{
\item{fit}{<obj> model object from which coefficients are extracted}
}
\description{
coefficients of a model fit
}
