% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_dymax}
\alias{metric_dymax}
\title{Maximum slope reached}
\usage{
metric_dymax(x, y, dy, ...)
}
\description{
x,y coordinates of first point of maximum dy. slope (m), y intercept (b),
x intercept (y0_x) of tangent line
}
