% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{metric_map}
\alias{metric_map}
\title{List of Performance Metrics}
\format{
A list where each item is a sublist containing the keys \emph{func} and
\emph{target}.
\describe{
\item{func}{The function used to calculate given metric.}
\item{target}{The value that would be reached in the case of optimal
performance.}
}
}
\usage{
metric_map
}
\description{
This list provides some common metrics of model performance along with
their "best value".
}
\keyword{datasets}
