% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_min_dbl}
\alias{metric_min_dbl}
\title{Minimum Doubling or Generation Time}
\usage{
metric_min_dbl(x, y, dy, ...)
}
\description{
Finds the point of maximum growth rate then treating that as the midpoint
of a doubling period, computes the doubling time at point.
}
