% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_year_strings}
\alias{parse_year_strings}
\title{Parse and generate lists of years.}
\usage{
parse_year_strings(year_strings)
}
\arguments{
\item{year_strings}{A vector of strings that each either represents a 
single year or a sequence of year in the format `start:stop`.}
}
\value{
run_years List of integer vectors, representing the years to 
  simulate for each run.
}
\description{
Parse and generate lists of years.
}
