% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pg_rosso}
\alias{pg_rosso}
\alias{pg_rosso_inv}
\title{Rosso Growth Model}
\usage{
pg_rosso(time, mu, lag, lower, alpha)

pg_rosso_inv(size, mu, lag, lower, alpha)
}
\arguments{
\item{time}{time}

\item{mu}{maximal growth rate}

\item{lag}{time lag}

\item{lower}{lower asymptote}

\item{alpha}{upper asymptote}

\item{size}{size (y-value)}
}
\description{
Rosso Growth Model

inverse of rosso model
}
\references{
T. Ross, "Indices for performance evaluation of predictive models in
food microbiology." J. Appl. Bacteriol. vol. 81, no. 5, pp. 501-508.
Nov. 1996.
}
