% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modvegesite.R
\name{plot.ModvegeSite}
\alias{plot.ModvegeSite}
\title{Plot ModVege simulation result overview}
\usage{
\method{plot}{ModvegeSite}(x, ...)
}
\arguments{
\item{x}{A \link{ModvegeSite} instance.}

\item{...}{Arguments are passed on to \link{ModvegeSite}\verb{$plot()}.}
}
\value{
NULL, but plots to the active device.
}
\description{
This wraps the \code{ModvegeSite} instance's \code{plot()} method.
}
\seealso{
The different \verb{[modvegeSite]$plot_XXX()} methods.
}
