% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_scan.R
\name{plot_parameter_scan}
\alias{plot_parameter_scan}
\title{Plot Parameter Scan Results}
\usage{
plot_parameter_scan(results)
}
\arguments{
\item{results}{Output of \code{\link[=analyze_parameter_scan]{analyze_parameter_scan()}}}
}
\description{
Visualize the results of a parameter scan and allow interactive inspection
of model performance in parameter space.
}
\details{
Under the hood this function just creates a PscanPlotter object and calls its
\code{analyze} method.
}
\seealso{
\code{\link[=analyze_parameter_scan]{analyze_parameter_scan()}}, \code{PscanPlotter$analyze()}
}
