% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\name{+.FunctionalGroup}
\alias{+.FunctionalGroup}
\title{Addition of two functional groups}
\usage{
\method{+}{FunctionalGroup}(A, B)
}
\arguments{
\item{A}{First functional group.}

\item{B}{Second functional group.}
}
\value{
C A \link{FunctionalGroup} object where each value is the sum of the
respective values in \emph{A} and \emph{B}.
}
\description{
Addition occurs by adding all FG parameters separately.
}
\examples{
fg1 = FunctionalGroup$new()
fg2 = FunctionalGroup$new(SLA = 0.02)
fg1 + fg2

}
\keyword{internal}
