% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess raw growth-curve measures}
\usage{
preprocess(y, log_base = exp(1), runmed_k = 3, runmav_n = 5,
  force_inc = TRUE, bg_subtract = function(y) y - min(y),
  calibrate_fxn = function(y) y)
}
\arguments{
\item{y}{\code{<num>} numeric vector of values to preprocess}

\item{log_base}{\code{<num>|<lgl>} base of log for log(y / min(y)) transformation.
Defaults to base \code{[exp(1)]}. set to \link{NA} or \link{FALSE} to skip.}

\item{runmed_k}{integer width of median window, must be odd. Passed to
\code{\link[stats:runmed]{stats::runmed()}}. Defaults to \code{3}. set to \link{NA} or \link{FALSE} to skip.}

\item{runmav_n}{integer width of moving average filter. Defaults to 5. set
to \link{NA} or \link{FALSE} to skip.}

\item{force_inc}{force y to be monotonically increasing?}

\item{bg_subtract}{numeric value to use for background substraction. Defaults
to \code{min(y)}. set to \link{NA} or \link{FALSE} to skip.}

\item{calibrate_fxn}{function to calibrate y values. Usually for correcting
for non-linear scattering at higher concentrations. Defaults to the
identity function \link{I}, that simply returns y un-transformed (as is).}
}
\value{
numeric vector of pre-processed values.
}
\description{
preprocess raw growth-curve measures
}
\examples{
y <- c(1:10, 8,11)
preprocess(y)
preprocess(1:10, log_base = F, runmed_k = F, runmav_n = F, force_inc = F, bg_subtract = F, calibrate_fxn = F)
}
