% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodvege-package.R
\docType{package}
\name{rmodvege-package}
\alias{rmodvege}
\alias{rmodvege-package}
\title{rmodvege: R implementation of the grass growth model ModVege.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Run grass growth simulations using a grass growth model based off of ModVege (Jouven, M., P. Carrère, and R. Baumont "Model Predicting Dynamics of Biomass, Structure and Digestibility of Herbage in Managed Permanent Pastures. 1. Model Description." (2006) \url{https://doi.org/10.1111/j.1365-2494.2006.00515.x}). See the project github page for a more thorough description: \url{https://github.com/kuadrat/rmodvege/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kuadrat/rmodvege}
  \item Report bugs at \url{https://github.com/kuadrat/rmodvege/issues/}
}

}
\author{
\strong{Maintainer}: Kevin Kramer \email{kevin.pasqual.kramer@protonmail.ch} (\href{https://orcid.org/0000-0001-5523-6924}{ORCID}) [copyright holder]

}
\keyword{internal}
