% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodvege-package.R
\docType{data}
\name{rmodvege_package_options}
\alias{rmodvege_package_options}
\title{Default options introduced by package rmodvege}
\format{
An object of class \code{list} of length 4.
}
\usage{
rmodvege_package_options
}
\description{
These are the default options, set when the package is loaded 
  by `library(rmodvege)`. To get or change the current value of an option, 
use the `options()` function.

\describe{
  \item{rmodvege.verbosity}{Integer that controls how much console output is 
        generated by rmodvege functions. Higher numbers mean more output. 
        See [logger()].}
  \item{rmodvege.input_dir}{Name of the directory in which to look for 
        input files.}
  \item{rmodvege.output_dir}{Name of the directory into which output files
        are written.}
  \item{rmodvege.data_dir}{Name of the directory in which to look for 
        measured data files.}
}
}
\seealso{
[options()]
}
\keyword{datasets}
