% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_loop_core}
\alias{run_loop_core}
\title{Core of growR_run_loop}
\usage{
run_loop_core(
  run_environment,
  run,
  n_runs,
  output_dir,
  return_results,
  independent
)
}
\arguments{
\item{run_environment}{A [ModvegeEnvironment] instance containing all the 
information necessary to simulate a site.}

\item{run}{integer Number of the currently carried out run.}

\item{n_runs}{integer Total number of runs to carry out.}

\item{output_dir}{str Path to directory where output files will be written 
to.}

\item{return_results}{boolean Whether or not simulation results are to be 
kept in memory and returned at function completion. Can be disabled to 
safe RAM, especially if results are anyways written to file (which is 
the case if `output_dir` is not empty).}

\item{independent}{boolean Whether or not simulations of subsequent years 
are treated as independent.}
}
\value{
A list of [ModvegeSite] instances containing the simulation 
  results for each year specified in *run_environment*.
}
\description{
Create the ModvegeSite object from supplied environment and call it's 
`$run()` method for each supplied year.
}
\seealso{
[growR_run_loop]
}
\keyword{internal}
