% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_loop_parallel}
\alias{run_loop_parallel}
\title{Run loop in parallel}
\usage{
run_loop_parallel(run_environment, ...)
}
\arguments{
\item{run_environment}{A ModvegeEnvironment instance; The environment with 
which to call [run_loop_core].}

\item{...}{Further arguments (except *run* and *n_runs*) to be passed to 
[run_loop_core].}
}
\value{
A list of [ModvegeSite] instances containing the simulation 
  results for each year specified in *run_environment*.
}
\description{
Simple wrapper around [run_loop_core].
}
\details{
Prints a message and passes dummy values for the *run* and *n_runs* 
variables, the former of which is not trivially available from within 
mclapply.
}
\seealso{
[run_loop_core]
}
\keyword{internal}
