% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{runmav}
\alias{runmav}
\title{running mean smoothing of length n}
\usage{
runmav(y, n)
}
\arguments{
\item{y}{values to process}

\item{n}{length of filter, width of window to apply mean smoothing to}
}
\description{
running mean smoothing of length n
}
