% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{SEA}
\alias{SEA}
\title{Seasonal effect on growth}
\usage{
SEA(ST, minSEA = 0.65, maxSEA = 1.35, ST1 = 800, ST2 = 1450)
}
\arguments{
\item{ST}{float Temperature sum in degree(C)-days.}

\item{minSEA}{float < 1. Minimum value of SEA.}

\item{maxSEA}{float > 1. Maximum value of SEA.}

\item{ST1}{float Temperature sum after which SEA declines from the maximum 
plateau.}

\item{ST2}{float Temperature sum after which SEA reaches and remains at 
its minimum.}
}
\description{
Function representing the strategy of plants adjusting their roots:shoots 
ratios during the season.
}
