% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\name{*.FunctionalGroup}
\alias{*.FunctionalGroup}
\title{Scalar multiplication of functional group (order matters: scalar * FG)}
\usage{
\method{*}{FunctionalGroup}(scalar, fg)
}
\arguments{
\item{scalar}{float to multiply all FG parameters with.}

\item{fg}{Functional group whose parameters are to be multiplied.}
}
\description{
Scalar multiplication of functional group (order matters: scalar * FG)
}
