% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{willmott}
\alias{willmott}
\title{Willmott Index}
\usage{
willmott(predicted, observed, ...)
}
\arguments{
\item{predicted}{Vector containing the predictions \emph{y}.}

\item{observed}{Vector containing the observations \emph{z}.}

\item{...}{Scaling factor \strong{c} in the denominator in the Willmott index.
The originally proposed value of 2 should be fine.}
}
\value{
willmott Value between -1 and 1
}
\description{
Willmott's index of model performance as described in
Willmott, Robeson and Matsuura: A refined index of model performance.
International Journal of Climatology 32: 2088-2094 (2012)
}
\details{
This index takes on values from -1 to 1, where values closer to 1 are
generally indicating better model performance. Values close to -1 can
either mean that the model predictions differ strongly from the
observation, or that the observations show small variance (or both).
}
\seealso{
\code{\link[=get_bias]{get_bias()}}
}
