% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{fW}
\alias{fW}
\title{Water stress}
\usage{
fW(W, PET)
}
\arguments{
\item{W}{Water stress given as the ratio of water reserves to water 
holding capacity.}

\item{PET}{Potential evapotranspiration in mm per day.}
}
\value{
A value in the range [0, 1], acting as a multiplicative factor to 
  plant growth.
}
\description{
Threshold function representing growth limitation due to water stress.
}
\details{
After equation (6) in McCall, D. G, und G. J Bishop-Hurley. A 
Pasture Growth Model for Use in a Whole-Farm Dairy Production Model, 
Agricultural Systems 76, Nr. 3 (1. Juni 2003): 1183 1205. 
https://doi.org/10.1016/S0308-521X(02)00104-X.
}
