% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{growR_run_loop}
\alias{growR_run_loop}
\title{Run a growR simulation}
\usage{
growR_run_loop(
  modvege_environments,
  write_files = TRUE,
  store_results = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{modvege_environments}{A list of \code{ModvegeEnvironment} instances.}

\item{write_files}{boolean; If TRUE, write results of \code{ModvegeSite$run()}
to files.}

\item{store_results}{boolean; If TRUE, return a list of the ModvegeSite`
objects which were run.}

\item{output_dir}{string; name of directory to which output files are to
be written if \code{write_files} is \code{TRUE}.}
}
\value{
If \code{store_results == TRUE}, a list of the format
\verb{[[run]][[year]]} containing clones of the \code{ModvegeSite} instances that
were run. Otherwise an empty list. Defaults to
getOption("growR.output_dir").
}
\description{
Start the loop over runs specified in the config file.
}
\details{
By default, returns an empty list but writes output to the output files
as specified in \code{modvege_environment}. Change this behaviour through the
\code{write_files} and \code{store_results} arguments.
}
\examples{
env1 = create_example_environment(site = "posieux")
env2 = create_example_environment(site = "sorens")

growR_run_loop(c(env1, env2), write_files = FALSE, 
store_results = TRUE)

}
