% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{management_parameters}
\alias{management_parameters}
\title{Management practices for Swiss grasslands}
\format{
A data.frame with 15 rows and 5 variables:
\describe{
\item{intensity}{Management intensity}
\item{altitude}{Altitude in m.a.s.l.}
\item{n_cuts}{Average number of cuts}
\item{yield}{Expected gross dry matter yield in t / ha}
\item{sigma_yield}{Uncertainty on yield in t / ha}
}
}
\usage{
management_parameters
}
\description{
Expected yields, uncertainties and average number of cuts as function of
altitude and management intensity.
}
\details{
Data after table 1b in

Olivier Huguenin et al.
Grundlagen für die Düngung landwirtschaftlicher Kulturen in der Schweiz
(GRUD), Kapitel 9: Düngung von Grasland
ISBN 1663-7852
https://www.agrarforschungschweiz.ch/2017/06/9-duengung-von-grasland-grud-2017/
}
\keyword{datasets}
