% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{modvege_run_loop}
\alias{modvege_run_loop}
\title{Run a rmodvege simulation}
\usage{
modvege_run_loop(
  modvege_environments,
  write_files = TRUE,
  store_results = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{modvege_environments}{A list of `ModvegeEnvironment` instances.}

\item{write_files}{boolean; If TRUE, write results of 
:class:`src.modvege_core.ModvegeSite` runs to files.}

\item{store_results}{boolean; If TRUE, return a list of the 
:class:`src.modvege_core.ModvegeSite` objects which were run.}

\item{output_dir}{string; name of directory to which output files are to 
be written if `write_files` is `TRUE`.}
}
\value{
If `store_results == TRUE`, a list of the format 
  `[[run]][[year]]` containing clones of the `ModvegeSite` instances that 
  were run. Otherwise an empty list. Defaults to 
  getOption("rmodvege.output_dir").
}
\description{
Start the loop over runs specified in the config file.
}
\details{
By default, returns an empty list but writes output to the output files 
as specified in `modvege_environment`. Change this behaviour through the 
`write_files` and `store_results` arguments.
}
