% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{ModvegeEnvironment}
\alias{ModvegeEnvironment}
\title{Data structure that contains inputs to rmodvege simulations}
\description{
Data structure that contains inputs to rmodvege simulations

Data structure that contains inputs to rmodvege simulations
}
\details{
This class contains site parameters, weather and management data for 
one simulation run of rmodvege on a given site over several years. 
Methods are provided to allow access to relevant data for a given year.

All inputs are read in from data files through the respective data classes
[WeatherData], [ManagementData] and [ModvegeParameters]. These parameters 
can be simultaneously specified through a config file.
}
\note{
Deprecated.
}
\seealso{
[WeatherData$get_weather_for_year()], 
  [ManagementData$get_management_for_year()]
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModvegeEnvironment-new}{\code{ModvegeEnvironment$new()}}
\item \href{#method-ModvegeEnvironment-load_inputs}{\code{ModvegeEnvironment$load_inputs()}}
\item \href{#method-ModvegeEnvironment-make_filename_for_run}{\code{ModvegeEnvironment$make_filename_for_run()}}
\item \href{#method-ModvegeEnvironment-get_environment_for_year}{\code{ModvegeEnvironment$get_environment_for_year()}}
\item \href{#method-ModvegeEnvironment-clone}{\code{ModvegeEnvironment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeEnvironment-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeEnvironment-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a new ModvegeEnvironment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeEnvironment$new(
  site_name,
  run_name = "-",
  years = "all",
  param_file = "-",
  weather_file = "-",
  management_file = "-",
  input_dir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site_name}}{string Name of the simulated site.}

\item{\code{run_name}}{string Name of the simulation run. Used to 
differentiate between different simulation conditions at the same site.
Defaults to "-", which indicates no specific run name.}

\item{\code{years}}{numeric Vector of integer years to be simulated.}

\item{\code{param_file}}{string Name of file that contains site and 
vegetation parameters. If default value `"-"` is  provided, it is 
assumed to be "SITENAME_parameters.csv".}

\item{\code{weather_file}}{string Analogous to *param_file*.}

\item{\code{management_file}}{string Analogous to *param_file*.}

\item{\code{input_dir}}{string Path to directory containing input files. 
Defaults to `getOption("rmodvege.input_dir")`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeEnvironment-load_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeEnvironment-load_inputs}{}}}
\subsection{Method \code{load_inputs()}}{
Load simulation inputs.

Stores parameters, management and weather data from files specified in 
self$parameter_file, self$weather_file and self$management_file, 
respectively.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeEnvironment$load_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeEnvironment-make_filename_for_run"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeEnvironment-make_filename_for_run}{}}}
\subsection{Method \code{make_filename_for_run()}}{
Ensure a readable filename for given *run_name*.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeEnvironment$make_filename_for_run(run_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run_name}}{Name of run to be converted into a filename.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeEnvironment-get_environment_for_year"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeEnvironment-get_environment_for_year}{}}}
\subsection{Method \code{get_environment_for_year()}}{
Get weather and environment for given year

Convenience function to retrieve environmental and management
inputs for given *year* from multi-year data containers `self$weather` 
and `self$management`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeEnvironment$get_environment_for_year(year)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{int; year for which to extract data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`list(W, M)` where `W` is the WeatherData and `M` the 
  ManagementData object for given year.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeEnvironment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeEnvironment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeEnvironment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
