% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modvegesite.R
\name{ModvegeSite}
\alias{ModvegeSite}
\title{ModvegeSite}
\description{
Implements the modvege grass growth model based off of Jouven et al. (2006).

This class contains model and site parameters and state variables as
attributes and has methods for running modvege with weather and management
input.

Use the \code{run()} method to carry out a simulation for a given year. The
results are stored in the state variables in this instance and can be written
to file using \code{write_output()}.
}
\note{
Technically, not all of the variables listed in
\code{state_variable_names} are re-initialized, as some of them already hold
the right values for the simulation. Particularly, this is the case for:
OMDDV, OMDDR
}
\section{Model variables}{
See Jouven et al. (2006) for a thorough description of all model variables.
\subsection{State Variables}{

\describe{
\item{BM}{Standing biomass in kg DM per ha.}
\item{BMG}{Standing green biomass (kg DM / ha).}
\item{cBM}{Cumulativeley grown biomass (kg DM / ha).}
\item{dBM}{Daily grown biomass (kg DM / ha).}
\item{hvBM}{Cumulative harvested biomass (kg DM / ha).}
\item{OMD}{Organic matter digestibility (kg / kg).}
\item{OMDG}{OMD of green matter (kg / kg).}
\item{ST}{Temperature sum in degree Celsius days.}
\item{REP}{Reproductive function. Gives the fraction of growth that is
assigned to reproductive growth. The remainder goes into vegetative growth.
Dimensionless.}
\item{PGRO}{Potential growth in kg DM / ha.}
\item{GRO}{Effective growth in kg DM / ha.}
\item{LAI}{Leaf area index, accounting for the proportion of light
intercepted by the sward. Dimensionless.}
\item{LAIGV}{LAI of green vegetative biomass. Dimensionless.}
\item{AET}{Actual evapotranspiration in mm.}
\item{WR}{Water reserves in mm.}
\item{ENV}{Function representing environmental effects on growth. Acts as a
multiplicative factor. Dimensionless.}
\item{ENVfPAR}{Part of ENV due to strength of incident radiation. Dimensionless.}
\item{ENVfT}{Part of ENV due to temperature. Dimensionless.}
\item{ENVfW}{Part of ENV due to water limitation. Dimensionless.}
}
}

\subsection{Initial conditions}{

\describe{
\item{AgeGV}{Age of green vegetative matter in degree Celsius days.}

\item{AgeGR}{Age of green reproductive matter in degree Celsius days.}

\item{AgeDV}{Age of dead vegetative matter in degree Celsius days.}

\item{AgeDR}{Age of dead reproductive matter in degree Celsius days.}

\item{BMGV}{biomass of GV (kg DM per ha).}

\item{BMGR}{biomass of GR (kg DM per ha).}

\item{BMDV}{biomass of DV (kg DM per ha).}

\item{BMDR}{biomass of DR (kg DM per ha).}

\item{BMDR}{biomass of DR (kg DM per ha).}

\item{SENGV}{senescence of GV (kg DM per ha).}

\item{SENGV}{senescence of GR (kg DM per ha).}

\item{ABSGV}{abscission of DV (kg DM per ha).}

\item{ABSGV}{abscission of DR (kg DM per ha).}

\item{ST}{thermal time (degree days).}

\item{cBM}{cumulative total biomass (kg per ha).}

}
}
}

\references{
{
\insertRef{jouven2006ModelPredictingDynamics}{rmodvege}
}
}
\seealso{
:func:\code{get_target_biomass}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time_step}}{Used time step in the model in days (untested).}

\item{\code{state_variable_namse}}{Vector containing the names of the model's
state variables.}

\item{\code{n_state_variables}}{Number of state variables.}

\item{\code{version}}{Version number of the rmodvege package. Is written into
output files.}

\item{\code{site_name}}{Name of the site to be simulated.}

\item{\code{run_name}}{Name of the simulation run. Used to distinguish between
different runs at the same site.}

\item{\code{year}}{Year to be simulated.}

\item{\code{j_start_of_growing_season}}{Index (DOY) of the day the growing season
was determined to begin.}

\item{\code{cut_height}}{Height of remaining grass after cut in m.}

\item{\code{parameters}}{A ModvegeParameters object.}

\item{\code{determine_cut}}{Function used to decide whether a cut occurs on a
given DOY. Is overloaded depending on whether management data is
provided or not.}

\item{\code{cut_DOYs}}{List of DOYs on which a cut occurred.}

\item{\code{cut_during_growth_preriod}}{Boolean to indicate whether a cut
occurred during the growth period, in which case reproductive growth is
stopped.}

\item{\code{last_DOY_for_initial_cut}}{\strong{autocut} Start cutting after this DOY,
even if yield target is not reached.}

\item{\code{max_cut_period}}{\strong{autocut} Maximum period to wait between
subsequent cuts.}

\item{\code{dry_precipitation_limit}}{\strong{autocut} Maximum amount of allowed
precipitation (mm) to consider a day.}

\item{\code{dry_days_before_cut}}{\strong{autocut} Number of days that shold be dry
before a cut is made.}

\item{\code{dry_days_after_cut}}{\strong{autocut} Number of days that shold be dry
after a cut is made.}

\item{\code{max_cut_delay}}{\strong{autocut} Number of days a farmer is willing to
wait for dry conditions before a cut is made anyways.}

\item{\code{cut_delays}}{\strong{autocut} Vector to keep track of cut delay times.
wait for dry conditions before a cut is made anyways.}

\item{\code{dry_window}}{\strong{autocut} Logical that indicates if DOY at index is
considered dry enough to cut.}

\item{\code{target_biomass}}{\strong{autocut} Biomass amount that should to be reached
by given DOY for a cut to be made.}

\item{\code{end_of_cutting_season}}{\strong{autocut} Determined DOY after which no
more cuts are made.}

\item{\code{BM_after_cut}}{\strong{autocut} Amount of biomass that remains after a cut
(determined through cut_height and biomass densities BDGV, BDDV, BDGR,
BDDR).}

\item{\code{weather}}{A WeatherData object.}

\item{\code{management}}{A ManagementData object. If its \code{is_empty} field is \code{TRUE},
the autocut routine will be employed.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{state_variable_namse}}{Vector containing the names of the model's
state variables.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModvegeSite-new}{\code{ModvegeSite$new()}}
\item \href{#method-ModvegeSite-initialize_state_variables}{\code{ModvegeSite$initialize_state_variables()}}
\item \href{#method-ModvegeSite-get_weather}{\code{ModvegeSite$get_weather()}}
\item \href{#method-ModvegeSite-get_management}{\code{ModvegeSite$get_management()}}
\item \href{#method-ModvegeSite-get_start_of_growing_season}{\code{ModvegeSite$get_start_of_growing_season()}}
\item \href{#method-ModvegeSite-carry_over_from_last_day}{\code{ModvegeSite$carry_over_from_last_day()}}
\item \href{#method-ModvegeSite-calculate_growth}{\code{ModvegeSite$calculate_growth()}}
\item \href{#method-ModvegeSite-calculate_ageing}{\code{ModvegeSite$calculate_ageing()}}
\item \href{#method-ModvegeSite-update_biomass}{\code{ModvegeSite$update_biomass()}}
\item \href{#method-ModvegeSite-calculate_digestibility}{\code{ModvegeSite$calculate_digestibility()}}
\item \href{#method-ModvegeSite-apply_cuts}{\code{ModvegeSite$apply_cuts()}}
\item \href{#method-ModvegeSite-determine_cut_from_input}{\code{ModvegeSite$determine_cut_from_input()}}
\item \href{#method-ModvegeSite-determine_cut_automatically}{\code{ModvegeSite$determine_cut_automatically()}}
\item \href{#method-ModvegeSite-check_dry_days}{\code{ModvegeSite$check_dry_days()}}
\item \href{#method-ModvegeSite-get_target_biomass}{\code{ModvegeSite$get_target_biomass()}}
\item \href{#method-ModvegeSite-run}{\code{ModvegeSite$run()}}
\item \href{#method-ModvegeSite-make_header}{\code{ModvegeSite$make_header()}}
\item \href{#method-ModvegeSite-write_output}{\code{ModvegeSite$write_output()}}
\item \href{#method-ModvegeSite-set_parameters}{\code{ModvegeSite$set_parameters()}}
\item \href{#method-ModvegeSite-clone}{\code{ModvegeSite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$new(parameters, site_name = "-", run_name = "-")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{A ModvegeParameters object.}

\item{\code{site_name}}{string Name of the simulated site.}

\item{\code{run_name}}{string Name of the simulation run. Used to
differentiate between different simulation conditions at the same site.
Defaults to "-", which indicates no specific run name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-initialize_state_variables"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-initialize_state_variables}{}}}
\subsection{Method \code{initialize_state_variables()}}{
Initialize all relevant state variables

Firstly, an empty vector of length \emph{days_per_year} is created for each
variable name in the vector \code{self$state_variable_names}.
Then, some additional variables relating to management and water balance
are also initialized.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$initialize_state_variables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-get_weather"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-get_weather}{}}}
\subsection{Method \code{get_weather()}}{
Return weather data if it exists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$get_weather()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-get_management"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-get_management}{}}}
\subsection{Method \code{get_management()}}{
Return management data if it exists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$get_management()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-get_start_of_growing_season"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-get_start_of_growing_season}{}}}
\subsection{Method \code{get_start_of_growing_season()}}{
Get index (or, equivalently, DOY) of the start of the growing season.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$get_start_of_growing_season(
  critical_temperature = 5,
  consider_snow = FALSE,
  critical_snow = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{critical_temperature}}{Daily average temperature in degree
Celsius required for a day being considered "warm enough" for growth.}

\item{\code{consider_snow}}{Toggle whether the effect of snow cover is to be
considered.}

\item{\code{critical_snow}}{Minimum daily snowfall in mm for a day to be
considered snowy.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-carry_over_from_last_day"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-carry_over_from_last_day}{}}}
\subsection{Method \code{carry_over_from_last_day()}}{
Set calculation values to the values stored in
previously used position.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$carry_over_from_last_day()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-calculate_growth"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-calculate_growth}{}}}
\subsection{Method \code{calculate_growth()}}{
Update all variables which change on a daily basis and
calculate biomass production.

This function updates the following attributes:
LAIGV
AET
WR
ENVfPAR
ENVfT
ENVfW
ENV
PGRO
GRO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$calculate_growth()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-calculate_ageing"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-calculate_ageing}{}}}
\subsection{Method \code{calculate_ageing()}}{
Calculate the ageing, senescence and abscission of different biomass
compartments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$calculate_ageing()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-update_biomass"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-update_biomass}{}}}
\subsection{Method \code{update_biomass()}}{
Calculate and update the amount of biomass in each
compartment.

:TODO: These derived values can be calculated vectorially upon request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$update_biomass()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-calculate_digestibility"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-calculate_digestibility}{}}}
\subsection{Method \code{calculate_digestibility()}}{
Caclulate the organic matter digestibility of each compartment.

:TODO: These derived values can be calculated vectorially upon
request and thus do not need to be part of a for loop.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$calculate_digestibility()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-apply_cuts"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-apply_cuts}{}}}
\subsection{Method \code{apply_cuts()}}{
Simulate the effects of cuts on the grass growth.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$apply_cuts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-determine_cut_from_input"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-determine_cut_from_input}{}}}
\subsection{Method \code{determine_cut_from_input()}}{
Read from the input whether a cut occurs on day \emph{j}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$determine_cut_from_input(DOY)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DOY}}{Integer day of the year for which to check.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-determine_cut_automatically"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-determine_cut_automatically}{}}}
\subsection{Method \code{determine_cut_automatically()}}{
Decide based on simple criteria whether day of year \emph{DOY} would be a
good day to cut.

This follows an implementation described in
Petersen, Krischan, David Kraus, Pierluigi Calanca, Mikhail A.
Semenov, Klaus Butterbach-Bahl, and Ralf Kiese. “Dynamic Simulation
of Management Events for Assessing Impacts of Climate Change on
Pre-Alpine Grassland Productivity.” European Journal of Agronomy
128 (August 1, 2021): 126306.
https://doi.org/10.1016/j.eja.2021.126306.

The decision to cut is made based on two criteria.
First, it is checked whether a \emph{target biomass} is reached on given
DOY. The defined target depends on the DOY and is given through
:func:\code{get_target_biomass}. If said biomass is present, return \code{TRUE}.

Otherwise, it is checked whether a given amount of time has passed
since the last cut. Depending on whether this is the first cut of
the season or not, the relevant parameters are
:int:\code{last_DOY_for_initial_cut} and :int:\code{max_cut_period}.
If that amount of time has passed, return \code{TRUE}, otherwise return
\code{FALSE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$determine_cut_automatically(DOY)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DOY}}{Integer day of the year for which to make a cut decision.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-check_dry_days"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-check_dry_days}{}}}
\subsection{Method \code{check_dry_days()}}{
Return TRUE if there are only \emph{dry} days in the next \emph{n}
days from given \emph{DOY}.

:TODO:
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$check_dry_days()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-get_target_biomass"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-get_target_biomass}{}}}
\subsection{Method \code{get_target_biomass()}}{
Get target value of biomass on given \emph{DOY}, which determines whether
a cut is to occur.

The regression for the target biomass is based on Fig. S2 in the
supplementary material of
Petersen, Krischan, David Kraus, Pierluigi Calanca, Mikhail A.
Semenov, Klaus Butterbach-Bahl, and Ralf Kiese. “Dynamic Simulation
of Management Events for Assessing Impacts of Climate Change on
Pre-Alpine Grassland Productivity.” European Journal of Agronomy
128 (August 1, 2021): 126306.
https://doi.org/10.1016/j.eja.2021.126306.

A refinement to expected yield as function of altitude has been
implemented according to Table 1a in
Huguenen-Elie et al. "Düngung von Grasland", Agrarforschung Schweiz,
8, (6), 2017,
https://www.agrarforschungschweiz.ch/2017/06/9-duengung-von-grasland-grud-2017/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$get_target_biomass(DOY, intensity = "high")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DOY}}{Integer day of the year to consider.}

\item{\code{intensity}}{One of {"high", "middle", "low"} specifying
management intensity..}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-run"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-run}{}}}
\subsection{Method \code{run()}}{
Carry out a ModVege simulation for one year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$run(year, weather, management)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{Integer specifying the year to consider. (Unused?)}

\item{\code{weather}}{Weather List for given year as returned by
:method:\code{Weather$get_weather_for_year}}

\item{\code{management}}{Management List for given year as provided by
:class:\code{ModvegeEnvironment}'s :method:\code{get_environment_for_year}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-make_header"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-make_header}{}}}
\subsection{Method \code{make_header()}}{
Construct a string full of meta information on a ModVege
run that can be inserted as a header to an output file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$make_header()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-write_output"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-write_output}{}}}
\subsection{Method \code{write_output()}}{
Write values of ModVege results into given file.

A header with metadata is prepended to the actual data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$write_output(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Path or name of filename to be created or overwritten.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-set_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-set_parameters}{}}}
\subsection{Method \code{set_parameters()}}{
Savely update the values in \code{self$parameters}.

Special care is taken to account for potential changes to
functional group weights.

:TODO: This could be implemented using just a list as input.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$set_parameters(param_names, param_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param_names}}{Vector containing the names of the parameters to be
updated.}

\item{\code{param_values}}{Vector containing the corresponding values to the
parameters in \emph{param_names}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModvegeSite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModvegeSite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModvegeSite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
