% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_group.R
\name{+.FunctionalGroup}
\alias{+.FunctionalGroup}
\title{Addition of two functional groups}
\usage{
\method{+}{FunctionalGroup}(A, B)
}
\arguments{
\item{A}{First functional group.}

\item{B}{Second functional group.}
}
\description{
Addition occurs by adding all FG parameters separately.
}
