% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posieux_weather}
\alias{posieux_weather}
\title{Example Weather Data}
\format{
A data.frame with 3652 rows and 10 variables:
  \describe{
    \item{year}{Year as an integer}
    \item{DOY}{Day of year as an integer}
    \item{Ta}{Average temperature of the day in degree Celsius}
    \item{Tmin}{Minimum temperature of the day in degree Celsius}
    \item{Tmax}{Maximum temperature of the day in degree Celsius}
    \item{precip}{Daily precipitation in mm}
    \item{rSSD}{Relative sunshine duration in percent}
    \item{SRad}{Sun irradiance in J / s / m^2. This can be converted into 
          photosynthetically active radiation (PAR) in MJ / m^2 as:
          PAR = SRad * 0.47 * 24 * 60 * 60 / 1e6}
    \item{ET0}{Evapotranspiration in mm.}
    \item{snow}{Precipitation in the form of snow in mm}
  }
}
\usage{
posieux_weather
}
\description{
Datasets containing the weather input parameters as used by rmodvege. The 
same data is made available as plain text files by the package and 
automatically found in the `input` directory created by [setup_directory()] 
if the `include_examples` option is set to `TRUE` (default).
}
\details{
For use in rmodvege, a WeatherData object has to be created from a plain 
text file. Therefore, this dataset is only provided for convenient 
inspection. In order to run rmodvege, use the plain text files provided by 
the package. Use `system.file("extdata", package = "rmodvege")` to locate them.

The *snow* column is not actually used by rmodvege but rather calculated 
through precipitation and temperatures in [WeatherData$read_weather()].

Likewise, the *rSSD* column is deprecated, currently unused and only kept 
for backwards compatibility.
}
\seealso{
[setup_directory()]
}
\concept{{weather_datasets}}
\keyword{datasets}
