% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodvege-package.R
\docType{package}
\name{rmodvege-package}
\alias{rmodvege}
\alias{rmodvege-package}
\title{rmodvege: R implementation of the grass growth model ModVege.}
\description{
The underlying grass growth model is based off of ModVege as described in Jouven, M., P. Carrère, and R. Baumont. “Model Predicting Dynamics of Biomass, Structure and Digestibility of Herbage in Managed Permanent Pastures. 1. Model Description.” Grass and Forage Science 61, no. 2 (2006): 112–24. https://doi.org/10.1111/j.1365-2494.2006.00515.x. It contains some expansions.
}
\author{
\strong{Maintainer}: Kevin Kramer \email{kevin.pasqual.kramer@protonmail.ch} (\href{https://orcid.org/0000-0001-5523-6924}{ORCID})

}
\keyword{internal}
