% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_rmodvege_verbosity}
\alias{set_rmodvege_verbosity}
\title{Set verbosity of rmodvege output.}
\usage{
set_rmodvege_verbosity(level)
}
\arguments{
\item{level}{Integer representing one of the following levels:
1: ERROR, 2: WARNING, 3: INFO, 4: DEBUG, 5: TRACE
Messages with a level higher than the specified *level* are suppressed.
In other words, higher values of *level* lead to more output and vice 
versa.}
}
\description{
Set verbosity of rmodvege output.
}
