% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_directory}
\alias{setup_directory}
\title{Initialize the desired directory structure for working with rmodvege.}
\usage{
setup_directory(root = ".", type = "empty")
}
\arguments{
\item{root}{Path to directory in which to initialize.}

\item{type}{Choose how things are to be initialized:
\itemize{
   \item "empty": Just put empty folders in place.
   \item "example": Fill the created folders with example data.
}}
}
\description{
Initialize the desired directory structure for working with rmodvege.
}
