% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_directory}
\alias{setup_directory}
\title{Initialize the desired directory structure for working with rmodvege.}
\usage{
setup_directory(root = ".", include_examples = TRUE)
}
\arguments{
\item{root}{Path to directory in which to initialize.}

\item{include_examples}{If `TRUE` (default), include example data and 
input parameters in the appropriate directories.}
}
\description{
Initialize the desired directory structure for working with rmodvege.
}
\seealso{
[examples]
}
