% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{WeatherData}
\alias{WeatherData}
\title{Weather Data Object}
\description{
Data structure containing weather data for a given site for several years.
}
\details{
All fields representing weather variables are vectors of length 365 * N, 
where *N* is the number of years for which weather data is stored. In 
other words, every variable has one value for each of the 365 of each of 
the *N* years.
}
\seealso{
[WeatherData$read_weather()]
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{years}}{numeric Integer representation of the contained years}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WeatherData-new}{\code{WeatherData$new()}}
\item \href{#method-WeatherData-read_weather}{\code{WeatherData$read_weather()}}
\item \href{#method-WeatherData-calculate_day_length}{\code{WeatherData$calculate_day_length()}}
\item \href{#method-WeatherData-get_weather_for_year}{\code{WeatherData$get_weather_for_year()}}
\item \href{#method-WeatherData-clone}{\code{WeatherData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WeatherData-new"></a>}}
\if{latex}{\out{\hypertarget{method-WeatherData-new}{}}}
\subsection{Method \code{new()}}{
Create a new WeatherData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WeatherData$new(weather_file = NULL, years = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weather_file}}{string Path to file containing the weather data to 
be read.}

\item{\code{years}}{numeric Vector of years for which the weather is to be 
extracted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WeatherData-read_weather"></a>}}
\if{latex}{\out{\hypertarget{method-WeatherData-read_weather}{}}}
\subsection{Method \code{read_weather()}}{
Read weather data from supplied *weather_file*.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WeatherData$read_weather(weather_file, years = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weather_file}}{Path to or name of file containing weather data.}

\item{\code{years}}{Years for which the weather is to be extracted.
Default is "all" to read in all found years.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WeatherData-calculate_day_length"></a>}}
\if{latex}{\out{\hypertarget{method-WeatherData-calculate_day_length}{}}}
\subsection{Method \code{calculate_day_length()}}{
Calculate the expected length of day based on a site's 
geographical latitude.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WeatherData$calculate_day_length(latitude)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{latitude}}{numeric; geographical latitude in degrees.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WeatherData-get_weather_for_year"></a>}}
\if{latex}{\out{\hypertarget{method-WeatherData-get_weather_for_year}{}}}
\subsection{Method \code{get_weather_for_year()}}{
Extract state variables to the weather data for given 
year and return them as a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WeatherData$get_weather_for_year(year)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{integer Year for which to extract weather data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WeatherData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WeatherData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WeatherData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
